Added the option to pass in a directory and recursively find subtitles.
See: [Issue #2: Add the option to recursively add subtitles to each video in a folder.](https://github.com/Tatsuonline/suboptimal/issues/2)

Co-authored-by: /u/ROFLLOLSTER <https://www.reddit.com/user/ROFLLOLSTER/>

Why: In case a user has a directory of multiple videos for which they want to get subtitles for, this change makes it very easy to accomplish.

How: A check was added to determine if the passed in argument is a file or directory. If it is a directory, it will recursively go through it and find subtitles for each file.

Tags: #recursion #directory #Reddit #ROFLLOLSTER #walkdir